requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer
Django==5.0.4
django-filter==24.2
django-rest-passwordreset==1.4.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.2.2
idna==3.7
pytz==2024.1
PyYAML==6.0
python-dotenv==1.0.0
requests==2.27.1
signals==0.0.2
sqlparse==0.5.0
typing_extensions==4.5.0
tzdata==2024.1
ujson==5.9.0
urllib3
Celery==5.2.7